Turns dbf tests module into a package
